Starter for `ArrayAccess` + more cleanup
